# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ApplyDiscountRequest:
    properties:
      code:
        example: WELCOME10
        maxLength: 32
        minLength: 1
        type: string
    required:
    - code
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CreateDiscountRequest:
    properties:
      active:
        description: Omitted defaults to TRUE.
        type: boolean
      code:
        example: WELCOME10
        maxLength: 32
        minLength: 3
        type: string
      description:
        maxLength: 2000
        type: string
      ends_at:
        type: string
      kind:
        enum:
        - percent
        - amount
        example: percent
        type: string
      max_discount:
        minimum: 1
        type: integer
      max_uses:
        minimum: 1
        type: integer
      min_order:
        minimum: 0
        type: integer
      starts_at:
        type: string
      value:
        example: 10
        minimum: 1
        type: integer
    required:
    - code
    - kind
    - value
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.DiscountAppliedResponse:
    properties:
      booking_id:
        type: string
      code:
        type: string
      discount:
        type: integer
      payable:
        type: integer
      subtotal:
        type: integer
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.DiscountCodeResponse:
    properties:
      active:
        type: boolean
      campaign_id:
        description: CampaignID is nil for a standalone code (PLAN_CAMPAIGN.md sections
          1-10).
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      ends_at:
        type: string
      id:
        type: string
      kind:
        type: string
      max_discount:
        type: integer
      max_uses:
        type: integer
      min_order:
        type: integer
      starts_at:
        type: string
      used_count:
        type: integer
      value:
        type: integer
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.UpdateDiscountRequest:
    properties:
      active:
        type: boolean
      description:
        maxLength: 2000
        type: string
      ends_at:
        type: string
      max_discount:
        minimum: 1
        type: integer
      max_uses:
        minimum: 1
        type: integer
      min_order:
        minimum: 0
        type: integer
      starts_at:
        type: string
      value:
        minimum: 1
        type: integer
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      details:
        additionalProperties:
          type: string
        type: object
      message:
        example: success
        type: string
      reason:
        description: |-
          Reason is a stable machine-readable key for an error ("discount_expired"), so the
          frontend can translate without matching on the English sentence. Absent on success.
        type: string
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Meta:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 42
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Paged:
    properties:
      items: {}
      meta:
        $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Meta'
    type: object
...
info:
  contact: {}
  description: API for the Cinema Project box-office system.
  title: BackEnd-CP API
  version: "1.0"
paths:
...
  /admin/discounts:
    get:
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Match code or description
        in: query
        name: search
        type: string
      - description: Filter enabled/disabled; omit for both
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Paged'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.DiscountCodeResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: List discount codes (admin)
      tags:
      - discounts
    post:
      consumes:
      - application/json
      parameters:
      - description: Code
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CreateDiscountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.DiscountCodeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: code already exists
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Create a discount code (admin)
      tags:
      - discounts
  /admin/discounts/{id}:
    delete:
      description: 'Soft delete: orders that used it keep their link, and the code
        string becomes reusable.'
      parameters:
      - description: Discount code id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Retire a discount code (admin)
      tags:
      - discounts
    get:
      parameters:
      - description: Discount code id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.DiscountCodeResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Get one discount code (admin)
      tags:
      - discounts
    patch:
      consumes:
      - application/json
      description: 'PARTIAL update. `code` and `kind` cannot be changed: either would
        rewrite what the code meant for orders that already used it.'
      parameters:
      - description: Discount code id
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.UpdateDiscountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.DiscountCodeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Update a discount code (admin)
      tags:
      - discounts
...
  /orders/{id}/discount:
    delete:
      description: Customer only. Gives the code's redemption back.
      parameters:
      - description: Booking id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.DiscountAppliedResponse'
              type: object
        "400":
          description: no discount to remove
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: no longer pending
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Remove the discount code from a pending order
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Customer only. `subtotal` stays the seat total; `payable` is what
        the gateway will charge. Every refusal is 400/40001 with a distinct message,
        and an unknown code answers the same message as a disabled one so the endpoint
        cannot be used to enumerate codes.
      parameters:
      - description: Booking id
        in: path
        name: id
        required: true
        type: string
      - description: Code
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ApplyDiscountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.DiscountAppliedResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: the order belongs to another user
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: already discounted, or no longer pending
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Apply a discount code to a pending order
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: 'Header format: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"